Project-with-vgg16/requirements.txt
torch
torchvision
opencv-python-headless
numpy
pillow
matplotlib
scikit-learn
tqdm
streamlit
tifffile
imagecodecs
